Fix for disksolve bug
Changing the solving portion of the code to be a function that is called by PeGSDiskSolve. For whatever reason, this fixes the bug where the solver stops solving all of the particles for images beyond the first one. This can be reproduced by simply copying the same test image 3 times in the directory and the resulting synthetic image is quite different image 2 and 3 in the series. To my knowledge this error appear in versions of matlab beyond 2022a.